Don't wait more than 20ms for a write
Fixes #1 (I forgot to solve it before...)

The less we leave it on disk the less time someone has to steal it.
In this SSD and NVME world this can probably be even less. 20ms is
apparently the write latency for a 5400rpm HDD.

We just get an empty string or EOF, which is also bad. This is probably
a FIXME but I don't think it matters ¯\_(ツ)_/¯