Raise watcher debounce timeout to 500ms
Basically the timeout must be as low as possible so we don't "leak"
secrets. But it also must be long enough for Github Actions agent (which
is superslow) to be able to write the JSON file in ours integration
tests to disk otherwise we get weird file problems like: Err` value: Os
{ code: 32, kind: Uncategorized, message: "The process cannot access the
file because it is being used by another process." } which are hard
to deal with (how do we know that the whole JSON file has been written?
poll it X times and check if we can parse it and have a timeout? ...)